requirements.txt
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
--no-deps xformers<0.0.27
--no-deps trl<0.9.0
peft
accelerate
bitsandbytes
PyYAML
datasets
transformers
torch
